+++ iParapheur/ThumbsViewController.m
@@ -277,7 +277,7 @@
 	CGSize size = [thumbCell maximumContentSize]; // Get the cell's maximum content size
 	NSInteger page = (showBookmarked ? [[bookmarked objectAtIndex:index] integerValue] : (index + 1));
 	
-	[thumbCell showText:[NSString stringWithFormat:@"%d", page]]; // Page number place holder
+	[thumbCell showText:[NSString stringWithFormat:@"%d", (int)page]]; // Page number place holder
 	[thumbCell showBookmark:[document.bookmarks containsIndex:page]]; // Show bookmarked status
 	
 	NSURL *fileURL = document.fileURL;
